Refactored distributions to new directory and added Normal.
